src/test/java/com/hyeonwoo/spring/jpa/springjpa/MemberRepositoryTest.java
package com.hyeonwoo.spring.jpa.springjpa;

import org.assertj.core.api.Assertions;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.Assert.*;
@RunWith(SpringRunner.class)
@SpringBootTest
public class MemberRepositoryTest {
    @Autowired MemberRepository memberRepository;

    @Test
    @Transactional // entityManager를 통한 데이터 변경은 항상 Transaction 안에서 실행되어야 한다.
    // Transactional 어노테이션은 테스트에서 실행 시 마지막에 롤백 시켜줌.
    @Rollback(value = false)
    public void testMember() throws Exception {
        // given
        Member member = new Member();
        member.setUsername("memberA");

        // when
        Long savedId = memberRepository.save(member);
        Member findMember = memberRepository.find(savedId);

        // then
        Assertions.assertThat(findMember.getId()).isEqualTo(member.getId());
        Assertions.assertThat(findMember.getUsername()).isEqualTo(member.getUsername());
        Assertions.assertThat(findMember).isEqualTo(member);
        System.out.println("findMember == member: =" + (findMember == member));
        // 같은 영속성 컨텍스트 안에서 id가 같을경우 같은 객체 반환

    }
}